+++ hotCheetos/src/com/kdars/HotCheetos/MapReduce/SimScoreMapper2.java
@@ -7,24 +7,24 @@ import java.util.Iterator;
 import java.util.Map;
 
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.MapWritable;
 import org.apache.hadoop.mapreduce.Mapper;
-import org.apache.hadoop.mapreduce.Mapper.Context;
 
 import com.kdars.HotCheetos.Config.Configurations;
 import com.kdars.HotCheetos.DB.DBManager;
 import com.kdars.HotCheetos.DocumentStructure.DocumentInfo;
 
-public class SimScoreMapper2  extends Mapper<IntWritable, MapWritable, IntWritable, MapWritable>{
+public class SimScoreMapper2  extends Mapper<LongWritable, MapWritable, IntWritable, MapWritable>{
 	
 	@Override
-	public void map(IntWritable docID, MapWritable termFreqMap, Context context) throws IOException, InterruptedException {
+	public void map(LongWritable docID, MapWritable termFreqMap, Context context) throws IOException, InterruptedException {
 		int docInfoMemoryLimit = Configurations.getInstance().getDocInfoListLimit();
 		int tableID = Configurations.getInstance().getTableID();
 		
-		DBManager.getInstance().deleteDuplicateScores(docID.get(), tableID);  //도중에 죽었다면 flag가 1로 되어있을 것이므로, flag가 1이면서 inputdocument id가 같은 row들을 db에서 지우고 다시 시작.
+		DBManager.getInstance().deleteDuplicateScores((int) docID.get(), tableID);  //도중에 죽었다면 flag가 1로 되어있을 것이므로, flag가 1이면서 inputdocument id가 같은 row들을 db에서 지우고 다시 시작.
 		
-		ArrayList<Integer> corpusDocIDList = DBManager.getInstance().flagInputAndGetCurrentDocIDsFromInvertedIndexTable(docID.get(), tableID);
+		ArrayList<Integer> corpusDocIDList = DBManager.getInstance().flagInputAndGetCurrentDocIDsFromInvertedIndexTable((int) docID.get(), tableID);
 		
 		while(corpusDocIDList.isEmpty()){  //만약에 input documents와 비교할 corpus document가 DB에 없다면 while문을 타지 않음.
 
@@ -46,17 +46,17 @@ public class SimScoreMapper2  extends Mapper<IntWritable, MapWritable, IntWritab
 		
 		}
 		
-		DBManager.getInstance().unflag_Score(docID.get(), tableID);  //score 계산 및 저장이 정상적으로 완료되었다면 flag를 0으로 다시 set해주기.
+		DBManager.getInstance().unflag_Score((int) docID.get(), tableID);  //score 계산 및 저장이 정상적으로 완료되었다면 flag를 0으로 다시 set해주기.
 		
 		return;
 	}
 	
-	private boolean simScore_Calculation_OneVSInputCorpus(IntWritable docID, MapWritable termFreqMap, ArrayList<DocumentInfo> corpusDocInfoList, int scoreTableID, int invertedIndexTableID){
+	private boolean simScore_Calculation_OneVSInputCorpus(LongWritable docID, MapWritable termFreqMap, ArrayList<DocumentInfo> corpusDocInfoList, int scoreTableID, int invertedIndexTableID){
 		StringBuilder csvContent = new StringBuilder();
 		int bulkInsertLimit = Configurations.getInstance().getbulkScoreLimit();
 		int bulkInsertLimitChecker = 0;
 
-		int docid1 = docID.get();
+		int docid1 = (int) docID.get();
 		for (DocumentInfo docInfo2 : corpusDocInfoList){
 			int docid2 = docInfo2.docID;
 			double simscore = calcSim(termFreqMap, docInfo2.termFreq);
